第一次修改

+++ src/main/java/com/sparkchain/dao/repository/ActivityRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author 张立明
  * @date 2019/2/20 9:04.
  */
-public interface ActivityRepository extends JpaRepository<Activity,Integer>{
+public interface ActivityRepository extends JpaRepository<Activity,Long>{
 }
